Move helper function to separate module.
