+++ src/normal.rs
@@ -11,6 +11,7 @@
  *
  */
 
+use crate::category::merge_multi_categories;
 use crate::underscore::replace_underscores;
 use crate::unicode::{casefold, normalize_nfkc};
 use regex::Regex;
@@ -112,30 +113,6 @@ fn replace_in_place(text: &mut String, regex: &Regex, replace_with: &str) {
     }
 }
 
-fn merge_multi_categories(text: &mut String) {
-    let mut indices = Vec::new();
-    let mut first = true;
-
-    // Find all colons except the last
-    for (idx, ch) in text.char_indices().rev() {
-        if ch != ':' {
-            continue;
-        }
-
-        if first {
-            first = false;
-            continue;
-        }
-
-        indices.push(idx);
-    }
-
-    // Replace all colons with dashes
-    for idx in indices {
-        text.replace_range(idx..idx + 1, "-");
-    }
-}
-
 #[test]
 fn test_normalize() {
     macro_rules! check {
@@ -229,26 +206,4 @@ fn test_normalize() {
         "protected::fragment::_template",
         "protected:fragment:_template",
     );
-}
-
-#[test]
-fn test_multi_category() {
-    macro_rules! check {
-        ($input:expr, $expected:expr $(,)?) => {{
-            let mut text = str!($input);
-            merge_multi_categories(&mut text);
-            assert_eq!(
-                text,
-                $expected,
-                "Merged multiple categories doesn't match expected",
-            );
-        }};
-    }
-
-    check!("", "");
-    check!("alpha", "alpha");
-    check!("alpha:beta", "alpha:beta");
-    check!("alpha:beta:gamma", "alpha-beta:gamma");
-    check!("alpha:beta:gamma:delta", "alpha-beta-gamma:delta");
-    check!("alpha:beta:gamma:delta:epsilon", "alpha-beta-gamma-delta:epsilon");
 }
